ecp/core/HttpManager: new singleton for better okHttpClient use
allows the reuse of the okHttpClient object, as recommended by the
library's documentation.

* okHttpClient reuse allows for HTTP/2 connecton reuse and reduces
overhead.

* sets eager timeout specs to circumvent Roku issues regarding a lack
of priority for the ECP web server and not handling requests promptly
enough when the appliance is busy, such as with fast browsing of
collections or search results or while gaming. A low timeout cancels
requests that are not being timely handled, eluding "catch up"
interactions that only generate out-of-sequence inputs.

* sets up a connection pool with a 1 sec keepalive so HTTP/2
connections can be reused and to try to keep the ECP web server from
sleeping.

* just for a tiny bit of performance, a quick request object cache is
put in place based on URL, for requests where lag is relevant (those
being key press/up/down requests).