ecp/request: key requests require url-encoding. Also fix for Search
url-encode the key parameter as it is going to be part of the request
URL, which is particularly important for the Lit_ case (for example,
when trying to send a Space).

All url-encoding calls, including that found in SearchRequest changed
to use StandardCharsets.UTF_8.name() or otherwise a MethodNotFound
exception was generated at runtime (maybe only on my development
Android device, but this is more compatible in general). Also,
UnsupportedEncodingException has to be handled, so we just omit the
encoding part inside the catch, expecting this exception to never
happen as UTF_8 should always be available for URL encoding.